Implemented addCoins result message
