+++ src/styles/modals.css
@@ -1,4 +1,4 @@
-/*--GENERAL-MODAL--*/
+/*--GENERIC-MODAL--*/
 
 .modal-screen {
 	position: fixed;
